# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Address:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: integer
      street:
        type: string
      zip:
        type: integer
    required:
    - city
    - country
    - street
    - zip
    type: object
...
  models.Tokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  models.User:
    properties:
      admin:
        type: boolean
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      invoiceAddress:
        $ref: '#/definitions/models.Address'
      lastName:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      shippingAddress:
        $ref: '#/definitions/models.Address'
    type: object
...
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  v1.userSignInInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  v1.userSignUpInput:
    properties:
      email:
        type: string
      login:
        type: string
      password:
        type: string
    required:
    - email
    - login
    - password
    type: object
  v1.userUpdateAddressInput:
    properties:
      invoiceAddress:
        $ref: '#/definitions/models.Address'
      shippingAddress:
        $ref: '#/definitions/models.Address'
    required:
    - invoiceAddress
    - shippingAddress
    type: object
  v1.userUpdateEmailInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  v1.userUpdatePasswordInput:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
host: localhost:8000
info:
  contact: {}
  description: API server
  title: FinlandShop API
  version: "0.5"
paths:
...
  /users/email:
    put:
      consumes:
      - application/json
      description: update current user email
      parameters:
      - description: email info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userUpdateEmailInput'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      summary: User update email
      tags:
      - users-auth
  /users/info:
    put:
      consumes:
      - application/json
      description: update current user address
      parameters:
      - description: address info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userUpdateAddressInput'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      summary: User update address
      tags:
      - users-auth
  /users/logout:
    post:
      consumes:
      - application/json
      description: logout current user by authentication header
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      summary: Logout current user
      tags:
      - users-auth
  /users/me:
    delete:
      consumes:
      - application/json
      description: Delete current user
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      summary: Delete current user
      tags:
      - users-auth
    get:
      consumes:
      - application/json
      description: get current user by authentication header
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      summary: Get current user
      tags:
      - users-auth
  /users/password:
    put:
      consumes:
      - application/json
      description: update current user password
      parameters:
      - description: password info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userUpdatePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      summary: User update password
      tags:
      - users-auth
  /users/refresh:
    post:
      consumes:
      - application/json
      description: user refresh tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: User Refresh Tokens
      tags:
      - users-auth
  /users/sign-in:
    post:
      consumes:
      - application/json
      description: login into user account
      parameters:
      - description: sign in info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userSignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: User SignIn
      tags:
      - users-auth
  /users/sign-up:
    post:
      consumes:
      - application/json
      description: create user account
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: User sign-up
      tags:
      - users-auth
securityDefinitions:
  AdminAuth:
    in: context
    name: Admin authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"